+++ wxt.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
     manifest_version: 3,
     name: '__MSG_extensionName__',
     short_name: '__MSG_extensionShortName__',
-    version: '5.1.4',
+    version: '5.1.6',
     description: '__MSG_extensionDescription__',
     default_locale: 'en',
     icons: {